add two players game
