+++ Tic-Tac-Toe/README.md
@@ -1,8 +1,8 @@
-# TODO'S
+# React + Vite
 
-1. For the current move only, show “You are at move #…” instead of a button.
-2. Rewrite Board to use two loops to make the squares instead of hardcoding them.
-3. Add a toggle button that lets you sort the moves in either ascending or descending order.
-4. When someone wins, highlight the three squares that caused the win (and when no one wins, display a message about the result being a draw).
-5. Display the location for each move in the format (row, col) in the move history list.
-6. build AI opponent
+This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+
+Currently, two official plugins are available:
+
+- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
+- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
