+++ Tic-Tac-Toe/src/App.jsx
@@ -1,0 +1,7 @@
+import Game from "./components/game";
+
+function App() {
+  return <Game />;
+}
+
+export default App;
