+++ Tic-Tac-Toe/src/components/square.jsx
@@ -1,0 +1,15 @@
+function Square({ value, onSquareClick, color }) {
+  return (
+    <button
+      className="square"
+      onClick={onSquareClick}
+      style={{
+        backgroundColor: color,
+      }}
+    >
+      {value}
+    </button>
+  );
+}
+
+export default Square;
